move flags to their own file in util
